RAG_Course/requirements.txt
streamlit
pandas
faiss-cpu
sentence-transformers
google-generativeai
numpy
python-dotenv
rank_bm25
pypdf
pymupdf
pytest
